pyproject.toml
[project]
name = "multi"
version = "0.1.0"
description = "Multi-model AI orchestration server"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    {name = "Multi-MCP Team"}
]

dependencies = [
    "fastmcp>=2.13",
    "litellm>=1.80",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0",
]

[project.scripts]
multi = "multi_mcp.cli:main"

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=9.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.15.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "pytest-xdist>=3.5.0",
    "pytest-recording>=0.13.4",
    "vcrpy>=7.0.0",
    "httpx>=0.26.0",

    # Linting & formatting
    "ruff>=0.14.0",
    "pyright>=1.1.407",

    # Optional tracing
    # "langsmith>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["multi_mcp*"]
exclude = ["tests*", "docs*", "ref*", "tmp*"]

[tool.setuptools.package-data]
multi_mcp = ["config/*.yaml", "prompts/*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]
norecursedirs = ["tmp", "ref", "docs", ".git", ".venv"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
    "--tb=short",
    "--color=yes",
    "--disable-recording",  # Disable VCR due to httpx/litellm compatibility issues
]
markers = [
    "e2e: End-to-end tests using Claude CLI (deselect with '-m \"not e2e\"')",
    "slow: Tests that take >5 seconds to execute",
    "integration: Integration tests",
    "unit: Unit tests",
]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.ruff]
line-length = 140
target-version = "py311"
exclude = ["tests/data/**", ".venv/**", "build/**", "dist/**", "ref/**", "tmp/**"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.isort]
known-first-party = ["multi_mcp"]

[tool.pyright]
include = ["multi_mcp"]
exclude = ["**/__pycache__", "**/.pytest_cache", "tests/data/"]
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
pythonVersion = "3.11"
pythonPlatform = "All"
